# backend/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpdNoteCreateRequest": {
        "properties": {
          "content": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "icd_code": {
            "anyOf": [
              {
                "maxLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icd Code"
          },
          "icd_description": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icd Description"
          },
          "note_type": {
            "enum": [
              "examination",
              "diagnosis",
              "plan",
              "general"
            ],
            "title": "Note Type",
            "type": "string"
          }
        },
        "required": [
          "note_type",
          "content"
        ],
        "title": "OpdNoteCreateRequest",
        "type": "object"
      },
      "OpdPrescriptionCreateRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OpdPrescriptionItemCreateRequest"
            },
            "maxItems": 30,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "items"
        ],
        "title": "OpdPrescriptionCreateRequest",
        "type": "object"
      },
      "OpdPrescriptionItemCreateRequest": {
        "properties": {
          "dosage": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Dosage",
            "type": "string"
          },
          "duration": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Duration",
            "type": "string"
          },
          "frequency": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Frequency",
            "type": "string"
          },
          "instructions": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "medicine_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medicine Id"
          },
          "medicine_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Medicine Name",
            "type": "string"
          },
          "quantity": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "route": {
            "anyOf": [
              {
                "enum": [
                  "oral",
                  "topical",
                  "iv",
                  "im",
                  "sc",
                  "inhalation",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "oral",
            "title": "Route"
          }
        },
        "required": [
          "medicine_name",
          "dosage",
          "frequency",
          "duration"
        ],
        "title": "OpdPrescriptionItemCreateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Hospital Management SaaS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/opd/visits/{visit_id}/notes": {
      "get": {
        "description": "List clinical notes for a visit.",
        "operationId": "list_notes_api_v1_opd_visits__visit_id__notes_get",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Visit Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "note_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "examination",
                    "diagnosis",
                    "plan",
                    "general"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Note Type"
            }
          }
        ],
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Notes Api V1 Opd Visits  Visit Id  Notes Get",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Notes",
        "tags": [
          "OPD"
        ]
      },
      "post": {
        "description": "Add clinical note to visit.",
        "operationId": "create_note_api_v1_opd_visits__visit_id__notes_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Visit Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpdNoteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Note Api V1 Opd Visits  Visit Id  Notes Post",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Note",
        "tags": [
          "OPD"
        ]
      }
    },
    "/api/v1/opd/visits/{visit_id}/prescriptions": {
      "get": {
        "description": "List prescriptions for a visit.",
        "operationId": "list_prescriptions_api_v1_opd_visits__visit_id__prescriptions_get",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Visit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Prescriptions Api V1 Opd Visits  Visit Id  Prescriptions Get",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Prescriptions",
        "tags": [
          "OPD"
        ]
      },
      "post": {
        "description": "Create e-prescription with line items.",
        "operationId": "create_prescription_api_v1_opd_visits__visit_id__prescriptions_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Visit Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpdPrescriptionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Prescription Api V1 Opd Visits  Visit Id  Prescriptions Post",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Prescription",
        "tags": [
          "OPD"
        ]
      }
    },
    "/api/v1/opd/visits/{visit_id}/prescriptions/{prescription_id}": {
      "get": {
        "description": "Get prescription detail with items.",
        "operationId": "get_prescription_api_v1_opd_visits__visit_id__prescriptions__prescription_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Visit Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "prescription_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Prescription Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Prescription Api V1 Opd Visits  Visit Id  Prescriptions  Prescription Id  Get",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Prescription",
        "tags": [
          "OPD"
        ]
      }
    },
...
  }
}